move MathematicalProgramSolverInterface to its own file
