+++ drake/solvers/CMakeLists.txt
@@ -81,8 +81,8 @@ drake_install_headers(
   constraint.h
   function.h
   mathematical_program.h
+  mathematical_program_solver_interface.h
   decision_variable.h
-  solution_result.h
   system_identification.h
   )
 drake_install_libraries(drakeOptimization)
