+++ drake/solvers/mosek_solver.h
@@ -5,7 +5,6 @@
 #include <Eigen/Core>
 
 #include "drake/solvers/mathematical_program.h"
-#include "drake/solvers/solution_result.h"
 
 namespace drake {
 namespace solvers {
