--- drake/solvers/solution_result.h
@@ -1,14 +1,0 @@
-#pragma once
-
-namespace drake {
-namespace solvers {
-
-enum SolutionResult {
-    kSolutionFound = 0,
-    kInvalidInput = -1,
-    kInfeasibleConstraints = -2,
-    kUnknownError = -3,
-  };
-
-}  // namespace solvers
-}  // namespace drake
